12/9/2017 Delivery
Project: Detect potential mining sites in DR Congo.
Solution: We created a user interface and QGIS script to automate the
process of extracting high-resolution satellite images. By modifying the
TensorFlow Object Detection API, we trained a Faster R-CNN model on the
training data we collected. We created a program to automatically test
images using our trained model and output a list of coordinates of
potential mining sites.
################
This interface automates the process of creating a list of
coordinates covering a specified area and the process of deleting
low-resolution images.
The QGIS script is able to extract high-resolution satellite images
corresponding to a certain area indicated by its coordinates.